changing qty with ajax

+++ app/Http/Controllers/PaiementController.php
@@ -7,6 +7,9 @@ use Stripe\Stripe;
 use Stripe\PaymentIntent;
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\Session;
+use App\Order;
+use DateTime;
 
 class PaiementController extends Controller
 {
@@ -38,12 +41,45 @@ class PaiementController extends Controller
 
   public function store(Request $request){
     
-    Cart::destroy();
 
+    $order = new Order();
     $data = $request->json()->all();
-    return $data["paymentIntent"];
+    $order->payment_intent_id = $data["paymentIntent"]["id"];
+    $order->amount = $data["paymentIntent"]["amount"];
+    $order->payment_created_at = (new DateTime())
+      ->setTimestamp($data["paymentIntent"]["created"])
+      ->format("Y-m-d H:i:s");
 
+    $products = [];
+    $i=0;
+
+    foreach(Cart::content() as $product){
+      $products["product_".$i][] = $product->model->title;
+      $products["product_".$i][] = $product->model->price;
+      $products["product_".$i][] = $product->qty;
+      $i++;
+    }
+
+    $order->products = serialize($products);
+
+    $order->user_id = 15;
+
+    $order->save();
+
+    if($data["paymentIntent"]["status"] == "succeeded"){
+      Cart::destroy();
+      Session::flash("success","Votre commande a été traitée avec succès.");
+      return response()->json(["success" => "Payment Intent Succeeded"]);
+     
+    }else{
+      return response()->json(["error" => "Payment Intent Not Succeeded"]);
+    }
+    
   }
 
+
+  public function merci(){
+    return Session::has("success") ? view("paiement.merci") : redirect()->route("boutique");
+  }
 
 }
